54 Methods with pointer receiver
We just saw two Abs methods. One on the *Vertex pointer type and the other on the MyFloat value type.

There are two reasons to use a pointer receiver. First, to avoid copying the value on each method call (more efficient if the value type is a large struct).
Second, so that the method can modify the value that its receiver points to.