57 Errors
An error is anything that can describe itself as an error string.
The idea is captured by the predefined, built-in interface type,
error, with its single method, Error, returning a string:

```
type error interface {
Error() string
}
```

The fmt package's various print routines automatically know to call the method when asked to print an error.